Add resource handler, index and 404 handlers
