+++ controllers/assessment.js
@@ -1,0 +1,1 @@
+module.exports = {};
