+++ views/404.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Page Not Found!</title>
+    </head>
+    <body>
+        404 Error: Nothing found!
+    </body>
+</html>
